+++ src/main/java/hudson/plugins/performance/parsers/ParserFactory.java
@@ -9,8 +9,11 @@ import java.io.IOException;
 import java.io.PrintStream;
 import java.util.Hashtable;
 import java.util.Map;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 public class ParserFactory {
+    private static final Logger LOGGER = Logger.getLogger(ParserFactory.class.getName());
 
     protected static final Map<String, String> defaultGlobPatterns = new Hashtable<String, String>();
 
@@ -25,24 +28,19 @@ public class ParserFactory {
     }
 
     public static PerformanceReportParser getParser(Run<?, ?> build, FilePath workspace, PrintStream logger, String glob, EnvVars env) throws IOException, InterruptedException {
-        if (defaultGlobPatterns.containsKey(glob)) {
-            return getParser(defaultGlobPatterns.get(glob), glob);
+        String expandGlob = env.expand(glob);
+        if (defaultGlobPatterns.containsKey(expandGlob)) {
+            return getParser(defaultGlobPatterns.get(expandGlob), expandGlob);
         }
 
-        String expandGlob = env.expand(glob);
-        try {
-            FilePath[] pathList = workspace.list(expandGlob);
-            for (FilePath src : pathList) {
-                // copy file (it can be on remote slave) to "../build/../temp/" folder
-                final File localReport = new File(build.getRootDir(), "/temp/" + src.getName());
-                if (src.isDirectory()) {
-                    logger.println("Performance: File '" + src.getName() + "' is a directory, not a Performance Report");
-                    continue;
-                }
-                src.copyTo(new FilePath(localReport));
-                return getParser(ParserDetector.detect(localReport.getPath()), glob);
-            }
-        } catch (IOException ignored) {
+        File path = new File(expandGlob);
+        return path.isAbsolute() ? getParserWithAbsolutePath(build, workspace, logger, path) : getParserWithRelativePath(build, workspace, logger, expandGlob);
+    }
+
+    private static PerformanceReportParser getParserWithRelativePath(Run<?, ?> build, FilePath workspace, PrintStream logger, String glob) throws IOException, InterruptedException {
+        PerformanceReportParser result = getParserUsingAntPatternRelativePath(build, workspace, logger, glob);
+        if (result != null) {
+            return result;
         }
 
         File report = new File(workspace.getRemote() + '/' + glob);
@@ -55,6 +53,74 @@ public class ParserFactory {
 
         return getParser(ParserDetector.detect(workspace.getRemote() + '/' + glob), workspace.getRemote() + '/' + glob);
     }
+
+    private static PerformanceReportParser getParserUsingAntPatternRelativePath(Run<?, ?> build, FilePath workspace, PrintStream logger, String glob) throws InterruptedException {
+        try {
+            FilePath[] pathList = workspace.list(glob);
+            for (FilePath src : pathList) {
+                // copy file (it can be on remote slave) to "../build/../temp/" folder
+                final File localReport = new File(build.getRootDir(), "/temp/" + src.getName());
+                if (src.isDirectory()) {
+                    logger.println("Performance: File '" + src.getName() + "' is a directory, not a Performance Report");
+                    continue;
+                }
+                src.copyTo(new FilePath(localReport));
+                return getParser(ParserDetector.detect(localReport.getPath()), glob);
+            }
+        } catch (IOException ignored) {
+            LOGGER.log(Level.FINE, "Cannot find report file using Ant pattern", ignored);
+        }
+        return null;
+    }
+
+
+    private static PerformanceReportParser getParserWithAbsolutePath(Run<?, ?> build, FilePath workspace, PrintStream logger, File path) throws IOException, InterruptedException {
+        PerformanceReportParser result = getParserUsingAntPatternAbsolutePath(build, workspace, logger, path);
+        if (result != null) {
+            return result;
+        }
+
+        if (!path.exists()) {
+            // if report on remote slave
+            FilePath localReport = new FilePath(new File(build.getRootDir(), "/temp/" + path.getName()));
+            localReport.copyFrom(new FilePath(workspace.getChannel(), path.getAbsolutePath()));
+            return getParser(ParserDetector.detect(localReport.getRemote()), path.getName());
+        }
+
+        return getParser(ParserDetector.detect(path.getAbsolutePath()), path.getAbsolutePath());
+    }
+
+    private static PerformanceReportParser getParserUsingAntPatternAbsolutePath(Run<?, ?> build, FilePath wsp, PrintStream logger, File path) throws InterruptedException {
+        try {
+            File parent = path.getParentFile();
+            FilePath workspace = new FilePath(wsp.getChannel(), parent.getAbsolutePath());
+            while (!workspace.exists()) {
+                parent = parent.getParentFile();
+                if (parent != null) {
+                    workspace = new FilePath(wsp.getChannel(), parent.getAbsolutePath());
+                } else {
+                    return null;
+                }
+            }
+
+            String glob = path.getAbsolutePath().substring(parent.getAbsolutePath().length() + 1);
+            FilePath[] pathList = workspace.list(glob);
+            for (FilePath src : pathList) {
+                // copy file (it can be on remote slave) to "../build/../temp/" folder
+                final File localReport = new File(build.getRootDir(), "/temp/" + src.getName());
+                if (src.isDirectory()) {
+                    logger.println("Performance: File '" + src.getName() + "' is a directory, not a Performance Report");
+                    continue;
+                }
+                src.copyTo(new FilePath(localReport));
+                return getParser(ParserDetector.detect(localReport.getPath()), localReport.getPath());
+            }
+        } catch (IOException ignored) {
+            LOGGER.log(Level.FINE, "Cannot find report file using Ant pattern", ignored);
+        }
+        return null;
+    }
+
 
     private static PerformanceReportParser getParser(String parserName, String glob) {
         if (parserName.equals(JMeterParser.class.getSimpleName())) {