+++ src/test/java/hudson/plugins/performance/parsers/ParserFactoryTest.java
@@ -26,29 +26,33 @@ public class ParserFactoryTest {
     @Test
     public void testFlow() throws Exception {
         EnvVars envVars = new EnvVars(new HashMap<String, String>());
-        FilePath path = new FilePath(new File("/"));
+        FreeStyleProject project = j.createFreeStyleProject();
+        FreeStyleBuildExt build = new FreeStyleBuildExt(project);
+
+        FilePath workspace = new FilePath(Files.createTempDir());
+        build.setWorkspace(workspace);
         String filePath;
 
         filePath = getClass().getResource("/TaurusXMLReport.xml").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof TaurusParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof TaurusParser);
 
         filePath = getClass().getResource("/JMeterResults.jtl").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof JMeterParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof JMeterParser);
 
         filePath = getClass().getResource("/TEST-JUnitResults.xml").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof JUnitParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof JUnitParser);
 
         filePath = getClass().getResource("/IagoResults.log").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof IagoParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof IagoParser);
 
         filePath = getClass().getResource("/WrkResultsQuick.wrk").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof WrkSummarizerParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof WrkSummarizerParser);
 
         filePath = getClass().getResource("/JMeterCsvResults.csv").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof JMeterCsvParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof JMeterCsvParser);
 
         filePath = getClass().getResource("/summary.log").toURI().getPath();
-        assertTrue(ParserFactory.getParser(null, path, null, filePath, envVars) instanceof JmeterSummarizerParser);
+        assertTrue(ParserFactory.getParser(build, workspace, null, filePath, envVars) instanceof JmeterSummarizerParser);
     }
 
     @Test
@@ -64,7 +68,7 @@ public class ParserFactoryTest {
     }
 
     @Test
-    @Issue("43503")
+    @Issue("JENKINS-43503")
     public void test43503() throws Exception {
         FreeStyleProject project = j.createFreeStyleProject();
         FreeStyleBuildExt build = new FreeStyleBuildExt(project);
@@ -86,6 +90,61 @@ public class ParserFactoryTest {
         String glob2 = "**/results/*.jtl";
         assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, glob2, envVars) instanceof JMeterParser);
     }
+
+
+    @Test
+    @Issue("JENKINS-45119")
+    public void test45119() throws Exception {
+        FreeStyleProject project = j.createFreeStyleProject();
+        FreeStyleBuildExt build = new FreeStyleBuildExt(project);
+        EnvVars envVars = new EnvVars(new HashMap<String, String>());
+
+        FilePath workspace = new FilePath(Files.createTempDir());
+        build.setWorkspace(workspace);
+
+        String absPath1 = getClass().getResource("/WrkResultsQuick.wrk").getPath();
+
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, absPath1, envVars) instanceof WrkSummarizerParser);
+
+        String absPath2 = getClass().getResource("/JMeterResults.jtl").getPath();
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, absPath2, envVars) instanceof JMeterParser);
+
+        FilePath results = workspace.child("results");
+        results.mkdirs();
+
+        FilePath child = results.child("result.wrk");
+        child.copyFrom(getClass().getResourceAsStream("/WrkResultsQuick.wrk"));
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, "results/result.wrk", envVars) instanceof WrkSummarizerParser);
+
+        FilePath child2 = results.child("result.jtl");
+        child2.copyFrom(getClass().getResourceAsStream("/JMeterResults.jtl"));
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, "results/result.jtl", envVars) instanceof JMeterParser);
+    }
+
+    @Test
+    @Issue("JENKINS-45119")
+    public void testAntAbsolutePath() throws Exception {
+        FreeStyleProject project = j.createFreeStyleProject();
+        FreeStyleBuildExt build = new FreeStyleBuildExt(project);
+        EnvVars envVars = new EnvVars(new HashMap<String, String>());
+
+        FilePath workspace = new FilePath(Files.createTempDir());
+        build.setWorkspace(workspace);
+
+        String path = getClass().getResource("/single_result/res.csv").getPath();
+        path = path.replace("res.", "*.");
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, path, envVars) instanceof JMeterCsvParser);
+
+        String path2 = getClass().getResource("/single_result/nested/res.jtl").getPath();
+        path2 = path2.replace("nested/res", "**/*");
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, path2, envVars) instanceof JMeterParser);
+
+        String path3 = getClass().getResource("/single_result/nested/res.jtl").getPath();
+        path3 = path3.replace("single_result", "**");
+        path3 = path3.replace("res.", "*.");
+        assertTrue(ParserFactory.getParser(build, build.getWorkspace(), null, path3, envVars) instanceof JMeterParser);
+    }
+
 
     public static class FreeStyleBuildExt extends FreeStyleBuild {